Moved zip utils to loaders-utils module
