+++ modules/i3s/test/index.js
@@ -8,8 +8,3 @@ import './i3s-attribute-loader.spec';
 import './i3s-content-loader.spec';
 import './i3s-building-scene-layer-loader.spec';
 import './arcgis-webscene-loader.spec';
-import './zip-utils/data-view-provider.spec';
-import './zip-utils/cd-file-header.spec';
-import './zip-utils/local-file-header.spec';
-import './search-from-the-end.spec';
-import './zip-utils/end-of-central-directory.spec';
