+++ modules/loader-utils/src/parse-zip/data-view-file-provider.ts
