+++ modules/loader-utils/src/parse-zip/end-of-central-directory.ts
