+++ modules/loader-utils/src/parse-zip/file-provider.ts
