+++ modules/loader-utils/src/parse-zip/local-file-header.ts
