+++ modules/loader-utils/test/zip-utils/search-from-the-end.spec.js
@@ -1,7 +1,7 @@
 import test from 'tape-promise/tape';
-import {DATA_ARRAY} from './data/test.zip.js';
-import {searchFromTheEnd} from '../src/lib/parsers/parse-zip/search-from-the-end';
-import {DataViewFileProvider} from '../src/lib/parsers/parse-zip/data-view-file-provider';
+import {DATA_ARRAY} from '@loaders.gl/i3s/test/data/test.zip.js';
+import {searchFromTheEnd} from '../../src/parse-zip/search-from-the-end';
+import {DataViewFileProvider} from '../../src/parse-zip/data-view-file-provider';
 
 test('SLPKLoader#searchFromTheEnd', async (t) => {
   t.equals(
