+++ modules/tile-converter/src/slpk-extractor/helpers/file-handle-provider.ts
@@ -1,4 +1,4 @@
-import {FileProvider} from '@loaders.gl/i3s';
+import {FileProvider} from '@loaders.gl/loader-utils';
 import {FileHandle} from './fs-promises';
 
 /**
